fix: include correct client in payment link UI
